Code style

+++ src/watch.coffee
@@ -4,6 +4,7 @@ each = require 'async-each'
 chokidar = require 'chokidar'
 debug = require('debug')('brunch:watch')
 sysPath = require 'path'
+{spawn} = require 'child_process'
 logger = require 'loggy'
 pushserve = require 'pushserve'
 # worker must be loaded before fs_utils
@@ -80,13 +81,13 @@ startServer = (config, callback = ->) ->
     else
       server.startServer port, publicPath, serverCb
   else if config.server.command
-    spawn = require('child_process').spawn
-    commandComponents = config.server.command.split " "
+    commandComponents = config.server.command.split ' '
     debug "Invoking custom server command with: #{config.server.command}"
-    throw new Error 'Server command needs to have at least one word' if(commandComponents.length == 0)
-    child = spawn(commandComponents.shift(), commandComponents, {stdio: 'inherit'});
-    # fn to kill the server
-    child.close= (cb)=>
+    unless commandComponents.length
+      throw new Error 'Custom server command invalid'
+    child = spawn commandComponents.shift(), commandComponents, stdio: 'inherit'
+    # fn to kill the custom server
+    child.close = (cb) =>
       child.kill()
       cb?()
